GCM mode: implement 4-bit table based GCM Galois 128-bit multiply
The key is converted into two smaller tables that have a 4-bit index,
one for the data high nibble, one for the data low nibble. So the key
data tables are 480 bytes, instead of the 4080 bytes size for the 8-bit
table.

Note, it still uses the same 512-byte const table (which can go into
Flash memory on an embedded system), indexed by 8-bit bytes, for the
multiply-by-256 operation.