GCM encryption tests using test vectors: fix failing tests
For plaintext with a length that is not a multiple of 16, the GHASH
input must have padding bytes set to zero.

I had read the spec as saying that the _plaintext_ padding bytes must be
set to zero (but that the trailing bytes of the GHASH input would be set
to the values of the AES encrypted counter output). But I guess I
misread it.