GCM multiply: some code enhancements to increase speed
At least, it increases speed in THUMB code for ARM that is compiled in "Optimise for size" mode.

On a 32-bit processor, using uint128_struct_t more in the table-driven implementations allows for speed-ups by handling 32 bits of data at a time, e.g. when doing block XOR.

Unroll loops a little to reduce time at start/end of loops.